+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove k8sclient.G8sClient
 - Remove versionbundle.Changelog
 - Remove github.com/giantswarm/cluster-api v0.3.13-gs